PageSettings: Show active weather layer in map title
Mirror the RASP field suffix for EDL altitude and XCTherm layer labels
in AppendOverlayTitle. Build pan-mode titles from the configured weather
page via MakeTitle, refresh the map-scale string in SendUIState, and
remove unused GetPanOverlayLabel helpers.
